requirements.txt
annotated-types
cachetools
certifi
charset-normalizer
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
httplib2
idna
Jinja2
MarkupSafe
numpy
opencv-python
pip
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic_core
pyparsing
python-dotenv
requests
rsa
tqdm
typing_extensions
uritemplate
urllib3
